Implement fetch (#6)
Implement a fetch method

Fetch method tries to get a value for a key. If a value is found, then it is returned.
If a value is not found, the block will be invoked and its return value
(given that it is not nil) will be written to the cache.