Optimize by reducing heap allocations
This commit optimizes the number of heap allocations performed in each
request. For example, `Array(UInt8).new(0)` allocates an array on the
heap whereas `Bytes.empty` involves no heap allocations — the `Slice` is
allocated on the stack with no buffer on the heap because it isn't
necessary to allocate an empty buffer.

Another (micro-)optimization used here is to replace the hash of opcodes
with a NamedTuple. Indexing into a NamedTuple happens at compile time so
there is no runtime performance cost at all.